[breaking change] Move CriterionEvaluation and CriterionOperator in BadgeUp.Types
