+++ BadgeUpClient/Types/CriterionEvaluation.cs
@@ -1,0 +1,9 @@
+﻿namespace BadgeUp.Types
+{
+	public class CriterionEvaluation
+	{
+		public string Type { get; set; }
+		public CriterionOperator Operator { get; set; }
+		public int Threshold { get; set; }
+	}
+}
